Create a drop in sensor replacement for the sensor
